redesigning course page
