+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import './css/main.css';
-import CatalogPage from './components/catalog/CatalogPage';
+import CatalogPage from './components/pages/CatalogPage';
 
 function App() {
   return <CatalogPage />;
