+++ src/components/catalog/CatalogFilter.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import '../../css/catalog/CatalogFilters.css';
+import '../../css/catalog/CatalogFilters.scss';
 
 export default class CatalogFilters extends Component {
   constructor(props) {
@@ -9,13 +9,13 @@ export default class CatalogFilters extends Component {
     this.modalContainer = React.createRef();
 
     this.state = {
-      filters: ['JavaScript', 'CSS', 'HTML', 'React', 'MongoDB', 'Express', 'Node']
+      filters: ['Beginner', 'Intermediate', 'Advanced'],
     }
   }
 
   // ===============================================================
   // Handlers
-    /**
+  /**
      * TODO: Handle when filter is clicked
      * @param {Object} selectedCourse the object containing the data for the course to be displayed
      */
@@ -29,12 +29,14 @@ export default class CatalogFilters extends Component {
 
   render() {
     return (
-      <div id="filter-item">
-        {this.state.filters.map(filter => (
-          <p className={filter}>
-            {filter}
-          </p>
-        ))}
+      <div id="filter-nav">
+        <div id="filter-item">
+          {this.state.filters.map(filter => (
+            <p className={filter}>
+              {filter}
+            </p>
+          ))}
+        </div>
       </div>
     );
   }
