--- src/css/catalog/CatalogCard.css
@@ -1,106 +1,0 @@
-#course-list {
-  flex-wrap: wrap;
-  color: var(--offwhite);
-  font-family: var(--fontfamily);
-}
-
-#course-item {
-  display: grid;
-  justify-content: space-evenly;
-  grid-template-columns: 90%;
-  padding: 14px;
-  outline: none;
-  width: 275px; 
-  height: auto;
-  margin: 15px;
-  text-align: center;
-  border-radius: 7px;
-  background-color: #454545;
-  box-shadow: 2px 2px 2px rgba(119, 117, 117, 0.2);
-}
-
-#course-item:active {
-  filter: brightness(75%);
-}
-
-/* #course-item:hover {
-  cursor: pointer;
-} */
-
-
-#list-all-courses {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  width: 100%;
-  flex-wrap: wrap;
-  color: var(--offwhite);
-  font-family: var(--fontfamily);
-}
-
-.course-title {    
-  font-size: 22px;
-  margin-top: 14px;
-  color: antiquewhite;
-}
-
-.course-progress{    
-  font-size: 14px;
-  margin-top: 6px;
-}
-
-.course-image{
-  width: 100%;
-}
-
-
-.loader {
-  text-align: center;
-}
-
-
-
-@media only screen and (max-width: 400px) {
-  #list-all-courses {
-    padding-top: 30px;
-  }
-}
-
-/* Both these media tags fixes the size of the mega menu when resized for small screens */
-@media only screen and (max-width: 600px) {
-  #list-all-courses {
-    padding-top: 20px;
-  }
-
-  #course-item {
-    padding-top: 15px;
-  }
-}
-
-/* Small devices (landscape phones, 576px and up) */
-@media (min-width: 576px) {
-  #course-item {
-    width: 50%;
-  }
-}
-
-/* Medium devices (tablets portrait, 768px and up)  */
-@media (min-width: 768px) {
-  #course-item {
-    width: 45%;
-  }
-}
-
-/* Large devices (desktops, 992px and up) */
-@media (min-width: 992px) {
-  #course-item {
-    width: 28.333%;
-  }
-}
-
-/* Extra large devices (large desktops, 1200px and up) */
-@media (min-width: 1200px) {
-  #course-item {
-    width: 23%;
-  }
-}
