+++ src/css/catalog/CatalogCard.scss
@@ -1,0 +1,45 @@
+#course-card {
+  // background-color: #0A090C;
+  background-color: #454545;
+  color: #EEEEEE;
+  height: 300px;
+  border-radius: 5px;
+  box-shadow: 2px 3px 0px 1px #640D14;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.course-image {
+  width: 150px;
+  margin: 10px;
+}
+
+/* Small devices (landscape phones, 576px and up) */
+/* @media (min-width: 576px) {
+  #course-item {
+    width: 50%;
+  }
+}
+
+/* Medium devices (tablets portrait, 768px and up)  */
+/* @media (min-width: 768px) {
+  #course-item {
+    width: 45%;
+  } 
+} */
+
+/* Large devices (desktops, 992px and up) */
+/* @media (min-width: 992px) {
+  #course-item {
+    width: 28.333%;
+  }
+} */
+
+/* Extra large devices (large desktops, 1200px and up) */
+/* @media (min-width: 1200px) {
+  #course-item {
+    width: 23%;
+  }
+}  */
