--- src/css/catalog/CatalogFilters.css
@@ -1,24 +1,0 @@
-#filter-item {
-  display: grid;
-  justify-content: space-evenly;
-  grid-template-columns: 90%;
-  padding: 14px;
-  outline: none;
-  width: 77%; 
-  height: auto;
-  margin: 15px;
-  text-align: center;
-  border-radius: 7px;
-  color: white;
-  background-color: #454545;
-  box-shadow: 2px 2px 2px rgba(119, 117, 117, 0.2);
-}
-
-#filter-item:active {
-  filter: brightness(75%);
-}
-
-#filter-item:hover {
-  cursor: pointer;
-}
-
