+++ src/css/catalog/CatalogFilters.scss
@@ -1,0 +1,31 @@
+#filter-nav {
+  // background-color: #0A090C;
+  background-color: #454545;
+  box-shadow: 2px 2px 2px rgba(119, 117, 117, 0.2);
+}
+
+#filter-item {
+  margin: auto;
+  max-width: 80%;
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-start;
+  align-content: space-between;
+  align-items: space-between;
+  padding: 15px 0px;
+  outline: none;
+  color: white;
+  
+  p {
+    padding-right: 12px;
+  }
+}
+
+#filter-item:active {
+  filter: brightness(75%);
+}
+
+#filter-item:hover {
+  cursor: pointer;
+}
+
