+++ src/css/catalog/CatalogList.scss
@@ -1,4 +1,14 @@
-#course-list {
+#list-all-courses {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  grid-template-rows: repeat(3, 1fr);
+  grid-column-gap: 30px;
+  grid-row-gap: 30px;
+  margin-top: 20px;
+}
+
+
+/* #course-list {
   flex-wrap: wrap;
   color: var(--offwhite);
   font-family: var(--fontfamily);
@@ -34,15 +44,15 @@
 
 .img-fluid {
   /* max-width: 100%; */
-  width: 275px; 
+  /* width: 275px; 
   height: 12rem;
   background-color: cadetblue;
   text-align: center;
   padding: 4rem;
   box-shadow: 2px 2px 2px rgba(119, 117, 117, 0.2);
-}
+} */
 
-.loader {
+/* .loader {
   text-align: center;
 }
 
@@ -50,10 +60,10 @@
   #list-all-courses {
     padding-top: 30px;
   }
-}
+}  */
 
 /* Both these media tags fixes the size of the mega menu when resized for small screens */
-@media only screen and (max-width: 600px) {
+/* @media only screen and (max-width: 600px) {
   #list-all-courses {
     padding-top: 20px;
   }
@@ -61,44 +71,44 @@
   #course-item {
     padding-top: 15px;
   }
-}
+} */
 
 /* Small devices (landscape phones, 576px and up) */
-@media (min-width: 576px) {
+/* @media (min-width: 576px) {
   #course-item {
     width: 50%;
     padding-left: 10px;
     padding-right: 10px;
     padding-top: 15px;
   }
-}
+} */
 
 /* Medium devices (tablets portrait, 768px and up)  */
-@media (min-width: 768px) {
+/* @media (min-width: 768px) {
   #course-item {
     width: 45%;
     padding-left: 15px;
     padding-right: 15px;
     padding-top: 15px;
   }
-}
+} */
 
 /* Large devices (desktops, 992px and up) */
-@media (min-width: 992px) {
+/* @media (min-width: 992px) {
   #course-item {
     width: 28.333%;
     padding-left: 15px;
     padding-right: 15px;
     padding-top: 15px;
   }
-}
+} */
 
 /* Extra large devices (large desktops, 1200px and up) */
-@media (min-width: 1200px) {
+/* @media (min-width: 1200px) {
   #course-item {
     width: 23%;
     padding-top: 15px;
     padding-left: 10px;
     padding-right: 10px;
   }
-}
+} */
