--- src/css/catalog/CatalogPage.css
@@ -1,38 +1,0 @@
-/* Mobile first approach. Responsive breakpoints are listed smallest to highest */
-/* Small devices, Phones up to 575px */
-body {
-  margin-top: 55px;
-  background-color: var(--darkgrey);
-  font-family: var(--fontfamily);
-  font-size: 20px;
-}
-
-.catalog-page{
-  display: grid;
-  grid-template-columns: 1fr 27rem;
-}
-
-.filterHeader, .coursesHeader{
-  display: grid;
-  justify-content: space-evenly;
-  grid-template-columns: 90%;
-  padding: 14px;
-  outline: none;
-  width: 77%; 
-  height: auto;
-  margin: 15px;
-  text-align: center;
-  border-radius: 7px;
-  color: white;
-  background-color: #454545;
-  box-shadow: 2px 2px 2px rgba(119, 117, 117, 0.2);
-}
-
-.coursesHeader{
-  margin: auto;
-  height: 6rem;
-  font-size: 37px;
-  padding-top: 29px;
-  margin-bottom: 1rem;
-}
-
