+++ src/css/main.css
@@ -11,7 +11,8 @@
 
 body {
   margin-top: var(--nav-height);
-  background-color: var(--wooloo);
+  background-color: #0A090C;
+  /* background-color: #454545; */
   font-family: var(--fontfamily);
   font-size: 20px;
   height: 100%;
