+++ src/css/pages/CatalogPage.scss
@@ -1,3 +1,30 @@
-#an-id {
-  outline: none;
+/* Mobile first approach. Responsive breakpoints are listed smallest to highest */
+/* Small devices, Phones up to 575px */
+
+.catalog-page {
+  #catalog-header {
+    color: #EEEEEE;
+    margin: 0.5em auto;
+    max-width: 80%;
+    display: flex;
+    
+    input {
+      align-content: flex-end;
+      align-items: flex-end;
+      border-radius: 5px;
+      background-color: #454545;
+      border: none;
+      font-size: 20px;
+
+      ::placeholder {
+        color: #EEEEEE;
+      }
+
+    }
+  }
+
+  .course-container {
+    margin: 0.5em auto;
+    max-width: 80%;
+  }
 }
